[feature] ArtifactSpoilerのWebアップデート機能
GitHubの指定したブランチのソースツリーが取得できるURLを利用して、
a_info.txt、k_info.txt、activation-info-table.c をダウンロードして
DBの情報とアーティファクトリストを自動的に更新する。
一度ダウンロードしたら得られたetagヘッダの内容を次回のチェックで
if-none-matchヘッダに設定する事で、一度起動したら更新があるまで
304 が帰ってくるので、その場合は更新をしないようにする。

また、ArtifactSpoiler.pyに多くのクラスを書きすぎているので、
クラス毎にモジュールに分割する。
